fix: u8 overflow in test image generators (CI debug build)
Test fixtures compute (v + 40) % 255 with v in 0..254 as u8 — overflows
in debug builds (overflow checks on), panicking on CI while local
release-mode runs wrapped. Cast to u32 before adding.

+++ src/dct.rs
@@ -813,7 +813,7 @@ mod stable_seed_tests {
         let (w, h) = (256u32, 256u32);
         let mut img = ImageBuffer::from_fn(w, h, |x, y| {
             let v = ((x * 3 + y * 7 + x * y) % 255) as u8;
-            Rgb([v, (v + 40) % 255, (v + 80) % 255])
+            Rgb([v, (v as u32 + 40) as u8 % 255, (v as u32 + 80) as u8 % 255])
         });
         let seed_before = stable_seed(&img);
